#1525 polyfill restructuring
